Add freeze layer argument
Added `freeze_layer` argument to set a layer freeze for hyperparameter tuning. This model is pre-trained model and has 11 layers. User may choose one of the numbers from 0 to 11 and the n-th and following layers will be set to trainable. The rest layers will be set to non-trainable. If user choose 0, then the module will go with all the parameters trainable.